added inheritance to increase code reuse
